Feature/automate model promotion (#10)
# Pull Request

## Checklist
- [x] Have you read Digital Catapult's [Code of
Conduct](https://github.com/digicatapult/.github/blob/main/CODE_OF_CONDUCT.md)?
- [x] I have performed a self-review of my own code.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [x] I have made corresponding changes to the documentation.
- [x] My changes generate no new warnings.
- [x] I have added tests that prove my fix is effective or that my
feature works.
- [x] New and existing unit tests pass locally with my changes.

## PR Type

Please delete options that are irrelevant.

- [x] Feature
- [x] Documentation Update


## Linked tickets

https://digicatapult.atlassian.net/browse/BRID-124

## High level description

<!-- One paragraph explanation of the feature. -->

## Detailed description

Use mlflow apis to automate the model registration with a specified name
and for adding an alias.

Add integration tests: For the model registration to work, it was
required to bring the mlflow up with a backend store (sqlite in this
case). Hence we updated the docker compose yaml file.


## Describe alternatives you've considered

<!-- A clear and concise description of the alternative solutions you've
considered. Be sure to explain why the existing customisability isn't
suitable for this feature. -->

## Operational impact

<!--- A description of any operational considerations associated with
the change. Is there anything in particular we should be looking at when
deploying the change to make sure it is working as intended. If
something goes wrong will any special actions be needed to revert the
change. -->

## Additional context

<!-- Add any other context or screenshots about the feature request
here. -->