fix file move issue (#14)
# Pull Request

## Checklist
- [x] Have you read Digital Catapult's [Code of
Conduct](https://github.com/digicatapult/.github/blob/main/CODE_OF_CONDUCT.md)?
- [x] I have performed a self-review of my own code.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have made corresponding changes to the documentation.
- [x] My changes generate no new warnings.
- [ ] I have added tests that prove my fix is effective or that my
feature works.
- [x] New and existing unit tests pass locally with my changes.

## PR Type


- [x] Bug Fix


## Linked tickets

<!-- If this PR is linked to a JIRA ticket or Github Issue, please
provide the ticket URL here. -->

## High level description

ensure safe move of file from source to destination. 

## Detailed description

Deals with `OSError: Invalid cross-device link` when one is on a local
disk and the other is on an external disk or a network-mounted volume).

Refer -
https://stackoverflow.com/questions/42392600/oserror-errno-18-invalid-cross-device-link

## Describe alternatives you've considered

<!-- A clear and concise description of the alternative solutions you've
considered. Be sure to explain why the existing customisability isn't
suitable for this feature. -->

## Operational impact

<!--- A description of any operational considerations associated with
the change. Is there anything in particular we should be looking at when
deploying the change to make sure it is working as intended. If
something goes wrong will any special actions be needed to revert the
change. -->

## Additional context

<!-- Add any other context or screenshots about the feature request
here. -->